use doubles for tie points

+++ s2tbx-rapideye-reader/src/main/java/org/esa/beam/dataio/rapideye/metadata/RapidEyeMetadata.java
@@ -340,15 +340,15 @@ public class RapidEyeMetadata extends XmlMetadata {
         return dataType.getDataType();
     }
 
-    public float[] getCornersLatitudes() {
-        float[] lats = null;
+    public double[] getCornersLatitudes() {
+        double[] lats = null;
         if (rootElement != null) {
             MetadataElement currentElement;
             if (((currentElement = rootElement.getElement(RapidEyeConstants.TAG_TARGET)) != null) &&
                     ((currentElement = currentElement.getElement(RapidEyeConstants.TAG_FOOTPRINT)) != null) &&
                     ((currentElement = currentElement.getElement(RapidEyeConstants.TAG_GEOGRAPHIC_LOCATION)) != null)) {
                 MetadataElement[] pointElements = currentElement.getElements();
-                lats = new float[4];
+                lats = new double[4];
                 for (MetadataElement element : pointElements) {
                     try {
                         if (RapidEyeConstants.TAG_TOP_LEFT.equals(element.getName())) {
@@ -369,15 +369,15 @@ public class RapidEyeMetadata extends XmlMetadata {
         return lats;
     }
 
-    public float[] getCornersLongitudes() {
-        float[] longs = null;
+    public double[] getCornersLongitudes() {
+        double[] longs = null;
         if (rootElement != null) {
             MetadataElement currentElement;
             if (((currentElement = rootElement.getElement(RapidEyeConstants.TAG_TARGET)) != null) &&
                     ((currentElement = currentElement.getElement(RapidEyeConstants.TAG_FOOTPRINT)) != null) &&
                     ((currentElement = currentElement.getElement(RapidEyeConstants.TAG_GEOGRAPHIC_LOCATION)) != null)) {
                 MetadataElement[] pointElements = currentElement.getElements();
-                longs = new float[4];
+                longs = new double[4];
                 for (MetadataElement element : pointElements) {
                     try {
                         if (RapidEyeConstants.TAG_TOP_LEFT.equals(element.getName())) {
